refactor(YB-169): validation 은 별도 폴더로 분리
