+++ yellobook-api/src/main/java/com/yellobook/common/validation/validator/ExistProductValidator.java
@@ -1,6 +1,6 @@
-package com.yellobook.common.validator;
+package com.yellobook.common.validation.validator;
 
-import com.yellobook.common.annotation.ExistProduct;
+import com.yellobook.common.validation.annotation.ExistProduct;
 import com.yellobook.domain.inventory.service.InventoryQueryService;
 import com.yellobook.error.code.InventoryErrorCode;
 import jakarta.validation.ConstraintValidator;
